refactor(calendar): split calendar views into desktop/mobile/shared, rename Day/WeekView
Renames DailyView -> DayView and WeeklyView -> WeekView, and moves the
calendar component set from a flat components/calendar/ into
desktop/, mobile/, and shared/ subfolders (DayColumn moves to
shared/; MobileCalendarView is renamed to mobile/DayPortraitView to
disambiguate it from the upcoming landscape day view). Pure move and
rename -- no logic changes. Prepares the ground for the new
phone-landscape views on top of this structure.
