+++ frontend/app/(main)/page.module.scss
@@ -95,7 +95,7 @@
 
   // Must be a real flex column, not plain block flow: CalendarNavbar and the Day/Week view
   // below it are siblings, and the Day/Week view's own root (.outerContainer in
-  // DailyView.module.scss/WeeklyView.module.scss) sizes itself with `flex: 1 1 auto` to fill
+  // DayView.module.scss/WeekView.module.scss) sizes itself with `flex: 1 1 auto` to fill
   // whatever's left under CalendarNavbar -- that only resolves correctly against a flex
   // parent. Without this, block layout instead sized that root to 100% of *this whole box*
   // (double-counting CalendarNavbar's own height), and the overflow:hidden below silently
@@ -107,13 +107,13 @@
   display: flex;
   flex-direction: column;
 
-  // DailyView/WeeklyView each own their own bounded internal scroll region now, so this
+  // DayView/WeekView each own their own bounded internal scroll region now, so this
   // never needs to scroll itself -- letting it stay a scroll container (even via the
   // overflow-x-only auto-quirk) just chains leftover wheel scroll into it once the inner
   // view maxes out, dragging CalendarNavbar along with it.
   overflow: hidden;
 
-  // MobileCalendarView (WeekStrip/CalendarHeader/DayColumn) is meant to run edge-to-edge,
+  // DayPortraitView (WeekStrip/CalendarHeader/DayColumn) is meant to run edge-to-edge,
   // not sit inset inside the desktop 10px padding.
   @media (width <= $breakpoint-phone) {
     padding: 0;
