+++ frontend/app/components/calendar/MobileFab.tsx
@@ -6,8 +6,8 @@ interface MobileFabProps {
   onClick: () => void;
 }
 
-// Pinned bottom-right, rendered as a page-level sibling outside MobileCalendarView's swipe
-// region (see MobileCalendarView.tsx's DayColumn drag wrapper) so the drag gesture never
+// Pinned bottom-right, rendered as a page-level sibling outside DayPortraitView's swipe
+// region (see DayPortraitView.tsx's DayColumn drag wrapper) so the drag gesture never
 // swallows a tap intended for this button.
 const MobileFab: React.FC<MobileFabProps> = ({ onClick }) => {
   return (
