+++ frontend/styles/MainLayout.module.scss
@@ -10,7 +10,7 @@
 // overscroll-behavior-x isn't inherited, and per spec it must be set on the root element
 // (html) to control viewport edge-swipe navigation, not on body -- browsers that propagate it
 // from body are deviating from spec. Nothing on the page claims a horizontal touch-drag by
-// default (only WeekStrip's and MobileCalendarView's own drag="x" wrappers do, and only where
+// default (only WeekStrip's and DayPortraitView's own drag="x" wrappers do, and only where
 // their pointerdown handlers are actually attached) -- anywhere else, an unclaimed horizontal
 // swipe falls through to the browser's own edge-swipe-to-navigate gesture, which visually
 // slides the whole page sideways as a back/forward-navigation preview. Disabling that here is
