+++ frontend/styles/components/atoms/BoxText.module.scss
@@ -99,7 +99,7 @@
   }
 
   // Stretch to fill the parent's height instead of the fixed Meeting Block height —
-  // used by WeeklyView, where the wrapping div's height already encodes the meeting's
+  // used by WeekView, where the wrapping div's height already encodes the meeting's
   // duration. Must come after &.meeting so it wins the height tie-break. TagList's
   // containerStyle override (BoxText.tsx) handles pinning tags under the time line
   // instead of the bottom for this case.
@@ -108,7 +108,7 @@
   }
 
   // Tightens spacing so title + time + tags all still fit in a fillHeight box shrunk to
-  // roughly half its normal height (DailyView: two meetings sharing one room row) --
+  // roughly half its normal height (DayView: two meetings sharing one room row) --
   // without this, tags get pushed past the box's bottom edge and clipped by overflow: hidden.
   &.compact {
     padding: 3px 5px;
