+++ frontend/styles/components/calendar/desktop/WeekView.module.scss
@@ -8,7 +8,7 @@
   font-family: 'Source Sans Pro', sans-serif;
 
   // Fills whatever's left under CalendarNavbar in .primaryCalendar (page.module.scss), not
-  // this element's full containing-block height -- see DailyView.module.scss's matching
+  // this element's full containing-block height -- see DayView.module.scss's matching
   // .outerContainer comment for the full explanation (same shared parent, same bug).
   flex: 1 1 auto;
   min-height: 0;
@@ -151,7 +151,7 @@
   left: 0;
   width: 100%;
   height: 2px;
-  background-color: $brand-pink-color; // Matching the DailyView color
+  background-color: $brand-pink-color; // Matching the DayView color
   z-index: 12; // Above meeting boxes (10), below the sticky .dayHeader (15) so it
                // doesn't draw over the day header when they overlap during scroll.
