+++ frontend/util/meetingFilters.ts
@@ -1,5 +1,5 @@
-// Shared calendar/mode-tag and room-key filtering logic used by both DailyView and
-// WeeklyView, so the two views can't drift apart on what a filter key or tag means.
+// Shared calendar/mode-tag and room-key filtering logic used by both DayView and
+// WeekView, so the two views can't drift apart on what a filter key or tag means.
 
 import { formatETDateString } from "./timeUtils";
 
@@ -69,7 +69,7 @@ interface DateRoomTagMeeting {
 
 /**
  * Filters a flat meeting list down to the ones visible on `date`, given the current room/tag
- * filters -- the exact date+room+tag matching WeeklyView.getMeetingsForDay uses per-day,
+ * filters -- the exact date+room+tag matching WeekView.getMeetingsForDay uses per-day,
  * extracted here so the mobile day view (which needs the same single-day filtering) can't
  * silently drift from it. A Hybrid meeting occupies both its physical room and its Zoom room,
  * so it stays visible if either resource's filter is enabled; Remote has neither -- it's gated
